pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.backends._legacy:_Backend"

[project]
name = "erasus-unlearn"
version = "0.1.0"
description = "Efficient Representative And Surgical Unlearning Selection — Universal Machine Unlearning via Coreset Selection"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.9"
authors = [
    {name = "Avaya Aggarwal", email = "aggarwal.avaya27@gmail.com" },
]
keywords = ["machine-unlearning", "coreset", "foundation-models", "privacy", "pytorch"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    "torch>=2.0",
    "numpy>=1.24",
    "Pillow>=9.0",
    "tqdm>=4.60",
    "pyyaml>=6.0",
    "transformers>=4.30",
]

[project.optional-dependencies]
full = [
    "diffusers>=0.20",
    "opacus>=1.3",
    "datasets>=2.14",
    "scikit-learn>=1.2",
    "matplotlib>=3.7",
    "seaborn>=0.12",
    "wandb>=0.15",
    "peft>=0.5",
]
dev = [
    "pytest>=7.4",
    "pytest-cov>=4.1",
    "ruff>=0.1",
    "mypy>=1.5",
    "pre-commit>=3.4",
]

[project.scripts]
erasus = "erasus.cli.main:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["erasus*"]

[tool.ruff]
line-length = 100
target-version = "py39"

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --tb=short"
